+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 1.  [#3522](https://github.com/influxdata/chronograf/pull/3522): Add support for Template Variables in Cell Titles
 1.  [#3559](https://github.com/influxdata/chronograf/pull/3559): Add ability to export and import dashboards
 1.  [#3556](https://github.com/influxdata/chronograf/pull/3556): Add ability to override template variables and time ranges via URL query
+1.  [#3814](https://github.com/influxdata/chronograf/pull/3814): Add pprof routes to chronograf server
 
 ### UI Improvements